ls_19 created Search component
