+++ store/src/app/administration/search/search.component.html
@@ -1,0 +1,1 @@
+<input type="text" (input)="searchInput($event)" class="search-input" placeholder="Search...">
