+++ store/src/app/administration/search/search.component.scss
@@ -1,0 +1,7 @@
+.search-input {
+  width: 480px;
+  height: 60px;
+  border-radius: 5px;
+  font-size: 32px;
+  line-height: 39px;
+}
